mnist_fullstack/backend/requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astunparse==1.6.3
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
fastapi==0.135.2
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
grpcio==1.80.0
h11==0.16.0
h5py==3.14.0
httptools==0.7.1
idna==3.11
keras==3.13.2
libclang==18.1.1
markdown-it-py==4.0.0
mdurl==0.1.2
ml-dtypes==0.5.4
namex==0.1.0
numpy==2.4.4
opt-einsum==3.4.0
optree==0.19.0
packaging==26.0
pillow==12.1.1
protobuf==7.34.1
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.20.0
python-dotenv==1.2.2
python-multipart==0.0.22
pyyaml==6.0.3
requests==2.33.1
rich==14.3.3
setuptools==82.0.1
six==1.17.0
starlette==1.0.0
tensorflow==2.21.0
termcolor==3.3.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
wheel==0.46.3
wrapt==2.1.2
